classifier/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
filelock==3.17.0
fsspec==2025.2.0
idna==3.10
Jinja2==3.1.6
lightning-utilities==0.13.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pylsl==1.16.2
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
scipy==1.15.2
six==1.17.0
soupsieve==2.6
sympy==1.13.1
torch==2.6.0
torch-hd==5.7.1
torchmetrics==1.6.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
